+++ docs/changelog.rst
@@ -5,6 +5,7 @@ Version 1.2
 ~~~~~~~~~~~
 
 * Fixed voice selection to use VoiceLocaleIdentifier on OS X instead of deprecated VoiceLanguage
+* Introduced speakToFile function to driver api to write speech direct to disk as audio file. Support for OS X only at this stage
 
 Version 1.1
 ~~~~~~~~~~~